+++ src/redux/selectors.js
@@ -1,0 +1,7 @@
+import {
+    get,
+} from 'lodash'
+
+export function getAlerts(state) {
+    return get(state, 'alerts')
+}