+++ CHANGELOG.md
@@ -20,6 +20,7 @@ the code was deployed.
 - Sentry integration (CU-d8qufg).
 - FontAwesome icons (CU-d8qufg).
 - React Navigation with a simple tab navigation bar (CU-d8qufg).
+- Redux for handling state (CU-d8qufg).
 
 ### Removed
 - `yarn.lock` (CU-d8qufg).